requirements.txt
streamlit
langchain
langchain-google-genai
langchain-chroma
pydantic
python-dotenv
aiofiles
langsmith

